finalisation de la partie server

+++ ShroomGoRestServer/src/main/java/fr/polytech/unice/si5/entity/Position.java
@@ -1,32 +1,32 @@
 package fr.polytech.unice.si5.entity;
 
 public class Position {
-    private float longitude;
-    private float latitude;
+    private Double longitude;
+    private Double latitude;
 
-    public Position(float longitude, float latitude){
+    public Position(Double longitude, Double latitude){
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
     @Override
     public String toString() {
-        return Float.toString(longitude) + "," + Float.toString(latitude);
+        return Double.toString(longitude) + "," + Double.toString(latitude);
     }
 }
